+++ src/commands/terminators/deno.ps1
@@ -6,4 +6,16 @@ $denoDir = $env:USERPROFILE + "\.deno"
 
 Remove-Item $denoDir -Recurse -Force
 
-Write-Host ">>! Done. PATH entries weren't removed (this is a TODO)."
+
+Write-Host ">>> Removing Deno from PATH..."
+Write-Host ">>| Just in case, a PATH backup will be made at $env:USERPROFILE\FKN_PATH_BACKUP-TERMINATE_DENO.txt."
+
+$oldPath = [System.Environment]::GetEnvironmentVariable("Path", "User")
+
+Out-File "$env:USERPROFILE\FKN_PATH_BACKUP-TERMINATE_DENO.txt"
+
+$newPath = ($oldPath.Split(";") | Where-Object { $_ -ne $env:USERPROFILE + "\.deno\bin" }) -join ";"
+
+[System.Environment]::SetEnvironmentVariable("Path", $newPath, "User")
+
+Write-Host ">>> Done."