+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Fixed Cmds properly ignoring double quotes but not single quotes.
 - Fixed Cmds removing the last character if only the first one was a quote.
 - Fixed some errors (especially interop/golang ones) being inconsistent or lacking info in the error message.
+- Fixed `terminate` not removing Deno and Bun from PATH.
 
 ## [5.0.0] (28-09-2025)